Fixed Game Dir Format
