+++ src/routes/gen-craft/+page.server.ts
@@ -1,4 +1,4 @@
-import { Game } from '../gen-craft/game';
+import { Game } from './Game/game';
 import type { PageServerLoad, Actions } from '../gen-craft/$types';
 
 export const load = ( async ({cookies}) => {
