+++ src/routes/gen-craft/Game/util.ts
@@ -1,0 +1,7 @@
+export function replaceNoneEmptyString(nestedList : string[][], replacement : string){
+    return nestedList.map(sublist =>
+        sublist.map(item =>
+            item !== "" ? replacement : item
+        )
+    )
+}
